--- tests/configs/ddrnet.yaml
@@ -1,37 +1,0 @@
-model:
-  name: ddrnet_segmentation
-  nodes:
-    - name: DDRNet
-    - name: DDRNetSegmentationHead
-      inputs: ["DDRNet"]
-      alias: "segmentation_head"
-      params:
-        attach_index: -1
-      losses:
-        - name: CrossEntropyLoss
-      metrics:
-        - name: JaccardIndex
-    - name: DDRNetSegmentationHead
-      inputs: ["DDRNet"]
-      alias: "aux_segmentation_head"
-      params:
-        attach_index: -2
-      remove_on_export: true
-      losses:
-        - name: CrossEntropyLoss
-
-trainer:
-  preprocessing:
-    train_image_size: [128, 128]
-    keep_aspect_ratio: false
-    normalize:
-      active: true
-
-  batch_size: 2
-  epochs: 1
-  n_workers: 8
-  validation_interval: 10
-  n_log_images: 8
-
-  callbacks:
-    - name: ExportOnTrainEnd
