+++ tests/configs/parking_lot.json
@@ -14,8 +14,8 @@
                 "shape": [
                     1,
                     3,
-                    128,
-                    160
+                    32,
+                    64
                 ],
                 "layout": "NCHW",
                 "preprocessing": {
@@ -42,8 +42,8 @@
                 "shape": [
                     1,
                     23,
-                    128,
-                    160
+                    32,
+                    64
                 ],
                 "layout": "NCHW"
             },
@@ -53,8 +53,8 @@
                 "shape": [
                     1,
                     4,
-                    128,
-                    160
+                    32,
+                    64
                 ],
                 "layout": "NCHW"
             },
@@ -64,7 +64,7 @@
                 "shape": [
                     1,
                     9,
-                    320
+                    32
                 ],
                 "layout": "NCD"
             },
@@ -74,7 +74,7 @@
                 "shape": [
                     1,
                     9,
-                    80
+                    8
                 ],
                 "layout": "NCD"
             },
@@ -84,7 +84,7 @@
                 "shape": [
                     1,
                     9,
-                    20
+                    2
                 ],
                 "layout": "NCD"
             },
@@ -94,10 +94,10 @@
                 "shape": [
                     1,
                     6,
-                    16,
-                    20
+                    4,
+                    8
                 ],
-                "layout": "NCHW"
+                "layout": "NCDE"
             },
             {
                 "name": "output1_yolov6r2",
@@ -105,10 +105,10 @@
                 "shape": [
                     1,
                     8,
-                    16,
-                    20
+                    4,
+                    8
                 ],
-                "layout": "NCHW"
+                "layout": "NCDE"
             },
             {
                 "name": "output2_yolov6",
@@ -116,10 +116,10 @@
                 "shape": [
                     1,
                     6,
-                    8,
-                    10
+                    2,
+                    4
                 ],
-                "layout": "NCHW"
+                "layout": "NCDE"
             },
             {
                 "name": "output2_yolov6r2",
@@ -127,8 +127,8 @@
                 "shape": [
                     1,
                     8,
-                    8,
-                    10
+                    2,
+                    4
                 ],
                 "layout": "NCDE"
             },
@@ -138,8 +138,8 @@
                 "shape": [
                     1,
                     6,
-                    4,
-                    5
+                    1,
+                    2
                 ],
                 "layout": "NCDE"
             },
@@ -149,8 +149,8 @@
                 "shape": [
                     1,
                     8,
-                    4,
-                    5
+                    1,
+                    2
                 ],
                 "layout": "NCDE"
             }
